requirements_scraper.txt
# Scraping
scrapy==2.11.0
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0
webdriver-manager==4.0.1
fake-useragent==1.4.0

# Database
supabase==2.11.0
gotrue==2.11.0
python-dotenv==1.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Utilities
pyyaml==6.0.1
python-dateutil==2.8.2
lxml==4.9.3